assets: refactor the asset server so that the asset directory isn't hard-coded
